Add option so that JSC will crash if it can't allocate executable memory for the JITs https://bugs.webkit.org/show_bug.cgi?id=123048 <rdar://problem/12856193>
Patch by Michael Saboff <msaboff@apple.com> on 2013-10-21
Reviewed by Geoffrey Garen.

Added new option, called crashIfCantAllocateJITMemory. If this option is true then we crash
when checking the validity of the executable allocator. The default value for this option is
false, but jsc sets it to true when built for iOS to make it straightforward to identify whether
the app can obtain executable memory.

* jsc.cpp: Explicitly enable crashIfCantAllocateJITMemory on iOS.
(main):
* runtime/Options.h: Added option crashIfCantAllocateJITMemory.
* runtime/VM.cpp:
(JSC::enableAssembler): Modified to crash if option crashIfCantAllocateJITMemory
is enabled.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@157719 268f45cc-cd09-0410-ab3c-d52691b4dbfc