Add CryptoKey base class and bindings https://bugs.webkit.org/show_bug.cgi?id=123216
Reviewed by Sam Weinig.

* crypto/CryptoKey.idl: Added.

* CMakeLists.txt:
* DerivedSources.make:
* GNUmakefile.list.am:
Process the IDL.

* WebCore.xcodeproj/project.pbxproj: Added files.

* bindings/js/JSCryptoKeyCustom.cpp: Added.
(WebCore::JSCryptoKey::algorithm): Use a visitor to build algorithm dictionary
for the key.

* crypto/CryptoAlgorithmDescriptionBuilder.cpp: Added.
* crypto/CryptoAlgorithmDescriptionBuilder.h: Added.
An interface for a visitor we'll use to expose CrytoKey.algorithm in bindings,
and possibly also for storage serialization. Not complete yet, we'll need support
for a few more simple types, and less trivially, for nested algorithms.

* bindings/js/JSCryptoAlgorithmBuilder.cpp: Added.
* bindings/js/JSCryptoAlgorithmBuilder.h: Added.
An implementation that builds an algorithm description dictionary for JS bindings.

* crypto/CryptoKey.cpp: Added.
(WebCore::CryptoKey::~CryptoKey):
(WebCore::CryptoKey::buildAlgorithmDescription):
* crypto/CryptoKey.h: Added.
Added an almost empty implementation. Some of the functions that are currently
marked as pure virtual will likely be implemented in this base class.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@157936 268f45cc-cd09-0410-ab3c-d52691b4dbfc