RenderElement::m_style should be a Ref. <https://webkit.org/b/123401>
Source/WebCore:

Made RenderElement::m_style a Ref. This codifies the fact that it
can never be null after construction.

Removed a couple of unnecessary null checks that were exposed as
compilation failures.

Reviewed by Antti Koivisto.

Source/WTF:

Added a Ref::replace() so we can Indiana Jones the new style in
RenderElement::setStyle() while keeping a handle on the old style
for a while longer.

Reviewed by Antti Koivisto.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@158111 268f45cc-cd09-0410-ab3c-d52691b4dbfc