Result of focus-option-control-on-page.html depends on system preferences https://bugs.webkit.org/show_bug.cgi?id=123304
Reviewed by Simon Fraser.

Source/WebKit/mac:

* WebView/WebView.mm: (-[WebView _retrieveKeyboardUIModeFromPreferences:]):
Read the preference from its correct domain. This doesn't affect behavior in practice
and isn't part of the fix, but it's a good opportunity to correct this code.

Source/WebKit2:

* UIProcess/mac/WKFullKeyboardAccessWatcher.mm:
(-[WKFullKeyboardAccessWatcher retrieveKeyboardUIModeFromPreferences:]):
Read the preference from its correct domain. This doesn't affect behavior in practice
and isn't part of the fix, but it's a good opportunity to correct this code.

Tools:

* DumpRenderTree/mac/DumpRenderTree.mm: (setDefaultsToConsistentValuesForTesting):
Override a preference that WebKit reads to decide what Tab does, so that current
user default doesn't affect tests.

* WebKitTestRunner/mac/main.mm: (setDefaultsToConsistentValuesForTesting):
Ditto.

LayoutTests:

* platform/mac/TestExpectations: Remove Pass/Failure expectations, this test
should always pass now.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@158256 268f45cc-cd09-0410-ab3c-d52691b4dbfc