Cleanup FontMac.mm https://bugs.webkit.org/show_bug.cgi?id=123928
Reviewed by Andy Estes.

* platform/graphics/mac/FontMac.mm:
(WebCore::showGlyphsWithAdvances): Inline the value of variable isVertical as we
reference it exactly once and its value is sufficiently clear.
(WebCore::Font::drawGlyphs): Remove default case in switch block so that the compiler
checks that all cases are covered. Move definition of platformData to the top of the
function and use it whenever we want to access the platform font data

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@158800 268f45cc-cd09-0410-ab3c-d52691b4dbfc