REGRESSION (r158014): Many webpages throw stack overflow exceptions on iOS (because Parser::parseMemberExpression uses ~130K more stack) https://bugs.webkit.org/show_bug.cgi?id=124177
Reviewed by Michael Saboff.

This patch pushes error handling into NEVER_INLINE functions to perform
the actual error message construction.  This dramatically reduces the
stack usage of the Parser.  For the large functions (such as parseMemberExpression)
the improvement is on the order of 2.5x reduction in stack usage.  For
smaller functions the reduction is in the order of 5-6x.

* parser/Parser.cpp:
(JSC::::logError):
* parser/Parser.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@159210 268f45cc-cd09-0410-ab3c-d52691b4dbfc