Add !USE(NETWORK_CFDATA_ARRAY_CALLBACK)-guard https://bugs.webkit.org/show_bug.cgi?id=124741
Reviewed by Alexey Proskuryakov.

Add !USE(NETWORK_CFDATA_ARRAY_CALLBACK)-guard around code that is unused
when building with feature NETWORK_CFDATA_ARRAY_CALLBACK.

Additionally, add a declaration for allocateSegment() with attribute WARN_UNUSED_RETURN
to have the compiler warn when the return value of this function is unused. Together with
warnings treated as errors this change will prevent a memory leak.

* platform/SharedBuffer.cpp:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@159650 268f45cc-cd09-0410-ab3c-d52691b4dbfc