[CSS Regions] Positioned elements in regions get clipped if they fall outside the region https://bugs.webkit.org/show_bug.cgi?id=117120
Reviewed by Mihnea Ovidenie.

Source/WebCore:

Fixed the computing of the box decorations clip rect when having statically positioned
elements inside positioned elements. The existing algorithm computed the rect in a loop
running up the containing block chain and only checked for positioned elements before
starting the loop. If a positioned elements was found in the middle of a loop (as would
be the case with a positioned element parenting a non-positioned element), it was not
correctly handled.
Also changed it so the clip is only performed in the fragmentation direction as that was
the original purpose of this method.

Tests: fast/regions/absolute-in-relative-overflow.html
       fast/regions/static-in-relative-overflow.html
       fast/regions/sticky-border-overflow.html

* rendering/RenderFlowThread.cpp:
(WebCore::RenderFlowThread::decorationsClipRectForBoxInRegion):

LayoutTests:

Added tests for the correct painting of the borders of differently positioned elements overflowing regions.

* fast/regions/absolute-in-relative-overflow-expected.html: Added.
* fast/regions/absolute-in-relative-overflow.html: Added.
* fast/regions/static-in-relative-overflow-expected.html: Added.
* fast/regions/static-in-relative-overflow.html: Added.
* fast/regions/sticky-border-overflow-expected.html: Added.
* fast/regions/sticky-border-overflow.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@160721 268f45cc-cd09-0410-ab3c-d52691b4dbfc