[GTK] Test /webkit2/WebKitWebResource/mime-type fails when run after /webkit2/WebKitWebView/resources https://bugs.webkit.org/show_bug.cgi?id=126119
Reviewed by Martin Robinson.

The problem is that when the blank.ico resource is loaded from the
disk cache, the mime type is null, because the soup cache doesn't
cache sniffed mime types. This doesn't happen when a resource is
loaded form the memory cache, because the ResourceResponse is
cached, not only the headers. I think the disk cache should also
cache the sniffed content type, but that needs to be done in
soup. For now we can workaround the issue in the unit test by
making sure that resources that can be cached, also include the
Content-Type header, this way the mime type won't be sniffed and
it will be cached as a HTTP header in the disk cache.

* UIProcess/API/gtk/tests/TestResources.cpp:
(serverCallback): Add Content-Type header for resources that can
be cached.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@160965 268f45cc-cd09-0410-ab3c-d52691b4dbfc