[GTK] Build with network process unconditionally https://bugs.webkit.org/show_bug.cgi?id=126128
Reviewed by Martin Robinson.

.:

* Source/autotools/SetupAutomake.m4: Remove network process
conditional for Makefiles.
* Source/autotools/SetupWebKitFeatures.m4: Remove network process
feature.

Source/WebKit2:

Build always with the network process enabled and decide whether
to use it or not using an environment variable. This makes a lot
easier to work on the network process and also to switch between
using it or not.

* GNUmakefile.am: Make sure we always include the WebKit2Prefix.h.
* UIProcess/API/gtk/WebKitWebContext.cpp:
(createDefaultWebContext): Use the network process when
WEBKIT_USE_NETWORK_PROCESS environment variable is present.
* WebKit2Prefix.h: Always enable network process for GTK port.

Tools:

* Scripts/webkitperl/FeatureList.pm: Remove network-process
feature command line option.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@160989 268f45cc-cd09-0410-ab3c-d52691b4dbfc