[CSS Regions] Fix painting when the composited region has overflow:hidden https://bugs.webkit.org/show_bug.cgi?id=124887
Patch by Mihai Maerean <mmaerean@adobe.com> on 2014-01-15
Reviewed by Alexandru Chiculita.

Source/WebCore:

When the layer of the region is composited, the region receives a GraphicsLayer of its own
so the clipping coordinates (caused by overflow:hidden) must be relative to the
GraphicsLayer coordinates in which the region gets painted.

Also, while the painting is done relative to the location of the region's content box, the
clipping is bound to the padding box of the region.

Tests: compositing/regions/paint-inside-composited-region-overflow-hidden-versus-div.html
       compositing/regions/paint-inside-composited-region-overflow-hidden-versus-region.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::paintFlowThreadIfRegion):

LayoutTests:

* compositing/regions/paint-inside-composited-region-overflow-hidden-versus-div.html: Added.
* compositing/regions/paint-inside-composited-region-overflow-hidden-versus-div-expected.html: Added.
* compositing/regions/paint-inside-composited-region-overflow-hidden-versus-region.html: Added.
* compositing/regions/paint-inside-composited-region-overflow-hidden-versus-region-expected.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@162115 268f45cc-cd09-0410-ab3c-d52691b4dbfc