[EFL][WK2] EwkView can not be shown without re-sizing once it is hidden. https://bugs.webkit.org/show_bug.cgi?id=127084
Reviewed by Gyuyoung Kim.

Visibility of EwkView can not be controlled by evas_object_show/hide
because EwkView can be shown only if re-sizing is requested, so modify
codes to show EwkView directly if we do not wait for re-sizing.
Additionally, we have to set default value of m_pendingSurfaceResize
to true for accelerated mode to prevent to show black empty view for
the first request to show.

* UIProcess/API/efl/EwkView.cpp:
(EwkView::EwkView):
(EwkView::handleEvasObjectShow):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@162182 268f45cc-cd09-0410-ab3c-d52691b4dbfc