Move didPerformFirstNavigation() logic to -[WebView _didCommitLoadForFrame:] https://bugs.webkit.org/show_bug.cgi?id=127189
Reviewed by Dan Bernstein.

Source/WebCore:

* loader/FrameLoaderClient.h:
(WebCore::FrameLoaderClient::didPerformFirstNavigation):
Add an empty implementation of didPerformFirstNavigation() to make it easier to remove it from subclasses.

Source/WebKit/mac:

This is a first step towards completely eliminating the didPerformFirstNavigation() client callback.

* WebCoreSupport/WebFrameLoaderClient.h:
* WebCoreSupport/WebFrameLoaderClient.mm:
Remove didPerformFirstNavigation() implementation.

* WebView/WebView.mm:
(-[WebView _checkDidPerformFirstNavigation]):
Check if we have both a current item and a previous item which indicates that we've navigated once.
The old code was called before there was a current item but since this is called from _didCommitLoadForFrame
we have to check that there's a current item as well.

(-[WebView _didCommitLoadForFrame:]):
Call _didCommitLoadForFrame.

* WebView/WebViewData.h:
Add new ivar.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@162224 268f45cc-cd09-0410-ab3c-d52691b4dbfc