[CSSRegions] Incorrect layout of a region pseudo children https://bugs.webkit.org/show_bug.cgi?id=126146
Reviewed by David Hyatt.

Source/WebCore:

Test: fast/regions/collapse-anonymous-region.html

A region behaviour, styled using -webkit-flow-from, is modeled using an anonymous
block created to fragment the named flow content inside the region. We have to prevent
the behaviour of anonymous children collapsing for this block to make sure that the
region element children are still laid out properly when the region element becomes an ordinary
element.

* rendering/RenderBlockFlow.h:

LayoutTests:

* fast/regions/collapse-anonymous-region-expected.html: Added.
* fast/regions/collapse-anonymous-region.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@162508 268f45cc-cd09-0410-ab3c-d52691b4dbfc