DebuggerCallFrame::evaluateWithCallFrame() should not execute a null executable. <https://webkit.org/b/127600>
Reviewed by Oliver Hunt.

In DebuggerCallFrame::evaluateWithCallFrame(), if the script string that
is passed in is bad, it will fail to create an Executable i.e.
EvalExecutable::create() returns a null pointer. However,
DebuggerCallFrame::evaluateWithCallFrame() was just clearing the
exception and proceeded to execute the null pointer as an Executable.
A crash ensues.

Now, if an exception is detected while creating the Executable, we
abort instead.

* debugger/DebuggerCallFrame.cpp:
(JSC::DebuggerCallFrame::evaluateWithCallFrame):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@162752 268f45cc-cd09-0410-ab3c-d52691b4dbfc