Teach Buildbot dashboard to parse alternative revision format https://bugs.webkit.org/show_bug.cgi?id=127561
Reviewed by Alexey Proskuryakov.

Add support for extracting the OpenSource and Internal revision numbers when Buildbot
returns a dictionary for the value of property got_revision.

* BuildSlaveSupport/build.webkit.org-config/public_html/dashboard/Scripts/BuildbotIteration.js:
(parseRevisionProperty): Added; parses Buildbot revision property, which can either be a
string or a dictionary.
(BuildbotIteration.prototype.update.): Modified to call parseRevisionProperty() to determine
the OpenSource and Internal revision numbers.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@162846 268f45cc-cd09-0410-ab3c-d52691b4dbfc